Distribution script fixes

+++ dist.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-set -x
+set -ex
 TMP=`mktemp -d`
 CWD=`pwd`
 
@@ -10,9 +10,10 @@ NEW_ARCHIVE_DIR="dist/"
 NEW_ARCHIVE_NAMES=("beacon-version-linux_x86_64.tar.gz" "beacon-version-linux_x86.tar.gz" "beacon-version-osx_x86_64.tar.gz" "beacon-version-win_x86_64.zip" "beacon-version-win_x86.zip")
 ADDITIONAL_FILES=("LICENSE.txt" "LICENSE_EXCEPTION.txt" "beacon.properties" "README")
 ECLIPSE_FILES=("eclipse-rcp-indigo-linux-gtk-x86_64.tar.gz" "eclipse-rcp-indigo-linux-gtk.tar.gz" "eclipse-rcp-indigo-macosx-cocoa-x86_64.tar.gz" "eclipse-rcp-indigo-win32-x86_64.zip" "eclipse-rcp-indigo-win32.zip")
-TUTORIAL_ARCHIVE_NAMES=("beacon-tutorial-version-linux_x86_64.tar.gz" "beacon-tutorial-version-linux_x86.tar.gz" "beacon-tutorial-version-osx_x86_64.tar.gz" "beacon-tutorial-version-win_x86_64.zip" "beacon-tutorial-version-win_x86.zip")
+TUTORIAL_ARCHIVE_NAMES=("beacon-tutorial-eclipse-version-linux_x86_64.tar.gz" "beacon-tutorial-eclipse-version-linux_x86.tar.gz" "beacon-tutorial-eclipse-version-osx_x86_64.tar.gz" "beacon-tutorial-eclipse-version-win_x86_64.zip" "beacon-tutorial-eclipse-version-win_x86.zip")
 TUTORIAL_ARCHIVE="beacon-tutorial-version.tar.gz"
 LOCAL_TARGET_ARCHIVE="beacon-local-target-version.tar.gz"
+TUTORIAL=0
 
 getExt() {
     mkdir -p ext
@@ -25,22 +26,30 @@ getExt() {
     cd ..
 }
 
-if [[ $# -lt 1 || $# -gt 2 ]]; then
-  echo "Usage: $0 <VERSION> <OpenFlowJ directory>"
+while getopts ":t" opt; do
+    case $opt in
+        t ) TUTORIAL=1 ;;
+        * )  ;;  
+    esac
+done
+shift $(($OPTIND - 1)) 
+
+if [[ $# -lt 2 ]]; then
+  echo "Usage: $0 [-t] <VERSION> <OpenFlowJ directory>"
   echo "       $0 1.0.0 ../openflowj"
   exit 65
 fi
 
 VERSION=$1
-if [[ "ext" == ${VERSION} ]]; then
-    getExt
-    exit 0
-fi
-
 OFJ=$2
 OFJ_VERSION=`grep -m 1 version ${OFJ}/pom.xml`
 if [[ $OFJ_VERSION =~ ^.*\<version\>([0-9.]*)\<\/version\>.*$ ]]; then
     OFJ_VERSION=${BASH_REMATCH[1]}
+    pushd .
+    cd ${OFJ}
+    mvn clean install
+    ./dist.sh
+    popd
 else
     echo "Failure determining OFJ version"
     exit 1
@@ -48,14 +57,15 @@ fi
 
 cd net.beaconcontroller.parent
 mvn clean
-mvn package
+mvn install
 cd ..
 
 # Delete any existing archives
 rm -rf ${CWD}/dist/*
 
 # Repack each archive with additional files listed above
-for i in `seq 0 ${#ARCHIVE_NAMES[@]}`; do
+MAXIT=`expr ${#ARCHIVE_NAMES[@]} - 1`
+for i in `seq 0 ${MAXIT}`; do
     ARCHIVE_NAME=${ARCHIVE_NAMES[$i]}
     # Extract existing archive
     if [[ $ARCHIVE_NAME =~ ^.*gz$ ]]; then
@@ -94,7 +104,6 @@ mkdir -p ${TMP}/beacon-${VERSION}
 cp -ar * ${TMP}/beacon-${VERSION}
 tar -czv -f ${CWD}/dist/beacon-${VERSION}-source.tar.gz --exclude-vcs --exclude="beacon-${VERSION}/dist" --exclude="beacon-${VERSION}/ext" --exclude="**/target" --exclude="**/bin" --exclude="**/logs" -C ${TMP} beacon-${VERSION}
 rm -rfd ${TMP}/*
-rm -rfd ${TMP}/.*
 
 cd net.beaconcontroller.parent
 mvn javadoc:javadoc
@@ -102,43 +111,49 @@ cd ..
 rm -rf dist/javadoc dist/apidocs
 cp -ar net.beaconcontroller.parent/target/site/apidocs dist/
 
-# Assemble the tutorial packages
-mkdir ${TMP}/beacon-tutorial-${VERSION}
-cp -ar ${CWD}/dist/apidocs ${TMP}/beacon-tutorial-${VERSION}
-mkdir ${TMP}/beacon-tutorial-${VERSION}/src
-tar xvzf ${CWD}/dist/beacon-${VERSION}-source.tar.gz -C ${TMP}/beacon-tutorial-${VERSION}/src
-tar xvzf ${OFJ}/dist/openflowj-${OFJ_VERSION}-source.tar.gz -C ${TMP}/beacon-tutorial-${VERSION}/src
-# extract the local target files here
-tar xzvf ${CWD}/ext/${LOCAL_TARGET_ARCHIVE//version/$VERSION} -C ${TMP}/beacon-tutorial-${VERSION}/src/beacon-${VERSION}
-# build the package without eclipse
-tar czvf ${CWD}/dist/${TUTORIAL_ARCHIVE//version/$VERSION} -C ${TMP} beacon-tutorial-${VERSION}
-# per eclipse version
-
-# Repack each archive with additional files listed above
-for i in `seq 0 ${#ECLIPSE_FILES[@]}`; do
-    ECLIPSE_FILE=${ECLIPSE_FILES[$i]}
-    TUT_ARCHIVE=${TUTORIAL_ARCHIVE_NAMES[$i]}
-    rm -rf ${TMP}/beacon-tutorial-${VERSION}/eclipse
-
-    # Extract the eclipse archive
-    if [[ $ECLIPSE_FILE =~ ^.*gz$ ]]; then
-        tar xzvf ${CWD}/ext/${ECLIPSE_FILE} -C ${TMP}/beacon-tutorial-${VERSION}
-    fi
-    if [[ $ECLIPSE_FILE =~ ^.*zip$ ]]; then
-        unzip ${CWD}/ext/${ECLIPSE_FILE} -d ${TMP}/beacon-tutorial-${VERSION}
-    fi
+if [[ $TUTORIAL -eq 1 ]]; then
+    echo "=== BUILDING TUTORIAL ==="
+    getExt
+    # Assemble the tutorial packages
+    mkdir ${TMP}/beacon-tutorial-${VERSION}
+    cp -ar ${CWD}/dist/apidocs ${TMP}/beacon-tutorial-${VERSION}
+    mkdir ${TMP}/beacon-tutorial-${VERSION}/src
+    tar xvzf ${CWD}/dist/beacon-${VERSION}-source.tar.gz -C ${TMP}/beacon-tutorial-${VERSION}/src
+    tar xvzf ${OFJ}/dist/openflowj-${OFJ_VERSION}-source.tar.gz -C ${TMP}/beacon-tutorial-${VERSION}/src
+    # extract the local target files here
+    tar xzvf ${CWD}/ext/${LOCAL_TARGET_ARCHIVE//version/$VERSION} -C ${TMP}/beacon-tutorial-${VERSION}/src/beacon-${VERSION}
+    # build the package without eclipse
+    tar czvf ${CWD}/dist/${TUTORIAL_ARCHIVE//version/$VERSION} -C ${TMP} beacon-tutorial-${VERSION}
+    # per eclipse version
     
-    # Pack archive
-    if [[ $TUT_ARCHIVE =~ ^.*gz$ ]]; then
-        tar czvf ${CWD}/dist/${TUT_ARCHIVE//version/$VERSION} -C ${TMP} beacon-tutorial-${VERSION}
-    fi
-    if [[ $TUT_ARCHIVE =~ ^.*zip$ ]]; then
-        pushd .
-        cd ${TMP}
-        zip -r ${CWD}/dist/${TUT_ARCHIVE//version/$VERSION} *
-        popd
-    fi
-done
+    # Repack each archive with additional files listed above
+    MAXIT=`expr ${#ECLIPSE_FILES[@]} - 1`
+    for i in `seq 0 ${MAXIT}`; do
+        ECLIPSE_FILE=${ECLIPSE_FILES[$i]}
+        TUT_ARCHIVE=${TUTORIAL_ARCHIVE_NAMES[$i]}
+        rm -rf ${TMP}/beacon-tutorial-${VERSION}/eclipse
+    
+        # Extract the eclipse archive
+        if [[ $ECLIPSE_FILE =~ ^.*gz$ ]]; then
+            tar xzvf ${CWD}/ext/${ECLIPSE_FILE} -C ${TMP}/beacon-tutorial-${VERSION}
+        fi
+        if [[ $ECLIPSE_FILE =~ ^.*zip$ ]]; then
+            unzip ${CWD}/ext/${ECLIPSE_FILE} -d ${TMP}/beacon-tutorial-${VERSION}
+        fi
+        
+        # Pack archive
+        if [[ $TUT_ARCHIVE =~ ^.*gz$ ]]; then
+            tar czvf ${CWD}/dist/${TUT_ARCHIVE//version/$VERSION} -C ${TMP} beacon-tutorial-${VERSION}
+        fi
+        if [[ $TUT_ARCHIVE =~ ^.*zip$ ]]; then
+            pushd .
+            cd ${TMP}
+            zip -r ${CWD}/dist/${TUT_ARCHIVE//version/$VERSION} *
+            popd
+        fi
+    done
+fi
 
 rm -rf ${TMP}
+echo "=== BUILD COMPLETE - FILES PLACED IN ${CWD}/dist ==="
